int cannot convert a float string
>>> s = '1.0'
>>> try:
...     int(s)
... except ValueError as e:
...     print(e)
invalid literal for int() with base 10: '1.0'
>>> int(float(s))
1

>>>

learned.

+++ sncompass/server.py
@@ -90,8 +90,8 @@ def compass(angle):
 
 @app.route('/crosshair/<x>/<y>/<z>')
 def crosshair(x, y, z):
-    svg_x = ( int(x) + 2000) / 100
-    svg_z = (-int(z) + 2000) / 100
+    svg_x = ( float(x) + 2000) / 100
+    svg_z = (-float(z) + 2000) / 100
 
     resp = make_response(
         render_template('SNMap-150.svg', x=svg_x, y=svg_z),
